feat:实现了小程序端套餐和菜品浏览的接口

+++ sky-server/src/main/java/com/sky/controller/admin/CategoryController.java
@@ -84,7 +84,7 @@ public class CategoryController {
    }
 
    @GetMapping("/list")
-   public Result<List> getCategoryByType(Long type){
+   public Result<List> getCategoryByType(Integer type){
 
         log.info("根据分类ID查询分类{}",type);
        List<Category> categoryList = categoryService.getCategoryByType(type);
